separate main_enclone_start; ./test passes
