+++ enclone_core/src/lib.rs
@@ -6,6 +6,7 @@ pub mod cell_color;
 pub mod combine_group_pics;
 pub mod convert_svg_to_png;
 pub mod defs;
+pub mod enclone_structs;
 pub mod join_one;
 pub mod linear_condition;
 pub mod logging;
