+++ enclone_stuff/src/lib.rs
@@ -4,8 +4,12 @@ pub mod disintegrate;
 pub mod doublets;
 pub mod fcell;
 pub mod filter_umi;
+pub mod flag_defective;
+pub mod inconsistent;
 pub mod merge_onesies;
 pub mod populate_features;
 pub mod some_filters;
 pub mod split_orbits;
+pub mod start;
+pub mod vars;
 pub mod weak_chains;
