+++ enclone_visual/src/enclone_server.rs
@@ -8,11 +8,12 @@ use crate::proto::{
 use crate::*;
 use chrono::prelude::*;
 use enclone_core::combine_group_pics::*;
+use enclone_core::enclone_structs::*;
 use enclone_core::logging::*;
 use enclone_core::parse_bsv;
 use enclone_main::main_enclone::*;
-use enclone_main::start::main_enclone_start;
 use enclone_main::stop::*;
+use enclone_stuff::start::main_enclone_start;
 use enclone_version::*;
 use flate2::write::GzEncoder;
 use flate2::Compression;
